# openapi.yaml
openapi: 3.1.0
info:
  title: Colette API
  description: Public REST API for the Colette app. Supports OAuth 2.0 and API key authentication.
  license:
    name: MIT
  version: 0.1.0
servers:
- url: /api
paths:
...
  /feedEntries:
    get:
      tags:
      - Feed Entries
      description: List feed entries
      operationId: listFeedEntries
      parameters:
      - name: feedId
        in: query
        description: Filter by the ID of the associated RSS feed
        required: false
        schema:
          type: string
          format: uuid
      - name: cursor
        in: query
        description: Pagination cursor
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Paginated list of feed entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paginated_FeedEntry'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /feedEntries/{id}:
    get:
      tags:
      - Feed Entries
      description: Get a feed entry by ID
      operationId: getFeedEntry
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Feed entry by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedEntry'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Feed entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /subscriptions:
    get:
      tags:
      - Subscriptions
      description: List user subscriptions
      operationId: listSubscriptions
      parameters:
      - name: filterByTags
        in: query
        description: Whether to filter by tags linked to the subscription
        required: false
        schema:
          type: boolean
      - name: tag[]
        in: query
        description: Filter by the IDs of the tags linked to the subscription
        required: false
        schema:
          type: array
          items:
            type: string
            format: uuid
      - name: withFeed
        in: query
        description: Whether to include the feed associated with the subscription
        required: false
        schema:
          type: boolean
      - name: withUnreadCount
        in: query
        description: Whether to include the count of the unread subscription entries associated with the subscription
        required: false
        schema:
          type: boolean
      - name: withTags
        in: query
        description: Whether to include the tags linked to the subscription
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Paginated list of subscriptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paginated_SubscriptionDetails'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - Subscriptions
      description: Subscribe to a web feed
      operationId: createSubscription
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionCreate'
        required: true
      responses:
        '201':
          description: Created subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Subscription already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /subscriptions/{id}:
    get:
      tags:
      - Subscriptions
      description: Get a subscription by ID
      operationId: getSubscription
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      - name: withFeed
        in: query
        description: Whether to include the feed associated with the subscription
        required: false
        schema:
          type: boolean
      - name: withUnreadCount
        in: query
        description: Whether to include the count of the unread subscription entries associated with the subscription
        required: false
        schema:
          type: boolean
      - name: withTags
        in: query
        description: Whether to include the tags linked to the subscription
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Subscription by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionDetails'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - Subscriptions
      description: Delete a subscription by ID
      operationId: deleteSubscription
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Successfully deleted subscription
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - Subscriptions
      description: Update a subscription by ID
      operationId: updateSubscription
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionUpdate'
        required: true
      responses:
        '200':
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /subscriptions/{id}/linkTags:
    patch:
      tags:
      - Subscriptions
      description: Link a list of tags to a subscription
      operationId: linkSubscriptionTags
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkSubscriptionTags'
        required: true
      responses:
        '204':
          description: Successfully linked tags
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      required:
      - code
      - message
      properties:
        code:
          $ref: '#/components/schemas/ApiErrorCode'
        message:
          type: string
    ApiErrorCode:
      type: string
      enum:
      - NOT_AUTHENTICATED
      - NOT_AUTHORIZED
      - NOT_FOUND
      - ALREADY_EXISTS
      - VALIDATION
      - BAD_GATEWAY
      - UNKNOWN
...
    Feed:
      type: object
      description: RSS feed
      required:
      - id
      - sourceUrl
      - link
      - title
      - description
      - refreshedAt
      - isCustom
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the feed
        sourceUrl:
          type: string
          format: uri
          description: URL to scrape for feed updates
        link:
          type: string
          format: uri
          description: URL of the webpage the feed links to
        title:
          type: string
          description: Title of the feed
        description:
          type:
          - string
          - 'null'
          description: Description of the feed
        refreshedAt:
          type:
          - string
          - 'null'
          format: date-time
          description: Timestamp at which the feed was refreshed
        isCustom:
          type: boolean
          description: Whether the feed was scraped from a custom plugin
...
    FeedEntry:
      type: object
      description: RSS feed entry
      required:
      - id
      - link
      - title
      - publishedAt
      - description
      - author
      - thumbnailUrl
      - feedId
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the feed entry
        link:
          type: string
          format: uri
          description: URL of the webpage the feed entry links to
        title:
          type: string
          description: Title of the feed entry
        publishedAt:
          type: string
          format: date-time
          description: Timestamp at which the feed entry was published
        description:
          type:
          - string
          - 'null'
          description: Description of the feed entry
        author:
          type:
          - string
          - 'null'
          description: Author of the feed entry
        thumbnailUrl:
          type:
          - string
          - 'null'
          format: uri
          description: Thumbnail URL of the feed entry
        feedId:
          type: string
          format: uuid
          description: Unique identifier of the associated RSS feed
...
    LinkSubscriptionTags:
      type: object
      description: Action to link tags to a user subscription
      required:
      - tagIds
      properties:
        tagIds:
          type: array
          items:
            type: string
            format: uuid
          description: Unique identifiers of the tags to link to the subscription
...
    Paginated_FeedEntry:
      type: object
      description: Paginated list of results
      required:
      - data
      properties:
        data:
          type: array
          items:
            type: object
            description: RSS feed entry
            required:
            - id
            - link
            - title
            - publishedAt
            - description
            - author
            - thumbnailUrl
            - feedId
            properties:
              id:
                type: string
                format: uuid
                description: Unique identifier of the feed entry
              link:
                type: string
                format: uri
                description: URL of the webpage the feed entry links to
              title:
                type: string
                description: Title of the feed entry
              publishedAt:
                type: string
                format: date-time
                description: Timestamp at which the feed entry was published
              description:
                type:
                - string
                - 'null'
                description: Description of the feed entry
              author:
                type:
                - string
                - 'null'
                description: Author of the feed entry
              thumbnailUrl:
                type:
                - string
                - 'null'
                format: uri
                description: Thumbnail URL of the feed entry
              feedId:
                type: string
                format: uuid
                description: Unique identifier of the associated RSS feed
          description: Current set of results
        cursor:
          type: string
          description: Pagination cursor, only present if more results are available
...
    Paginated_SubscriptionDetails:
      type: object
      description: Paginated list of results
      required:
      - data
      properties:
        data:
          type: array
          items:
            type: object
            description: Extended details of a user subscription
            required:
            - subscription
            properties:
              subscription:
                $ref: '#/components/schemas/Subscription'
                description: Subscription itself, always present
              feed:
                $ref: '#/components/schemas/Feed'
                description: Associated RSS feed, present if requested
              tags:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
                description: Linked tags, present if requested
              unreadCount:
                type: integer
                format: int64
                description: Count of unread subscription entries associated with the subscription, present if requested
          description: Current set of results
        cursor:
          type: string
          description: Pagination cursor, only present if more results are available
...
    Subscription:
      type: object
      description: User subscription to an RSS feed
      required:
      - id
      - title
      - description
      - feedId
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the subscription
        title:
          type: string
          description: Human-readable name of the subscription
        description:
          type:
          - string
          - 'null'
          description: Description of the subscription
        feedId:
          type: string
          format: uuid
          description: Unique identifier of the associated RSS feed
        createdAt:
          type: string
          format: date-time
          description: Timestamp at which the subscription was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp at which the subscription was modified
    SubscriptionCreate:
      type: object
      description: Data to create a new user subscription
      required:
      - title
      - feedId
      properties:
        title:
          type: string
          description: Human-readable name for the new subscription, cannot be empty
          minLength: 1
        description:
          type:
          - string
          - 'null'
          description: Description for the new subscription, cannot be empty
          minLength: 1
        feedId:
          type: string
          format: uuid
          description: Unique identifier of the associated RSS feed
    SubscriptionDetails:
      type: object
      description: Extended details of a user subscription
      required:
      - subscription
      properties:
        subscription:
          $ref: '#/components/schemas/Subscription'
          description: Subscription itself, always present
        feed:
          $ref: '#/components/schemas/Feed'
          description: Associated RSS feed, present if requested
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          description: Linked tags, present if requested
        unreadCount:
          type: integer
          format: int64
          description: Count of unread subscription entries associated with the subscription, present if requested
...
    SubscriptionUpdate:
      type: object
      description: Updates to make to an existing subscription
      properties:
        title:
          type:
          - string
          - 'null'
          description: Human-readable name for the subscription to update, cannot be empty
          minLength: 1
        description:
          type:
          - string
          - 'null'
          description: Description for the subscription to update, cannot be empty
          minLength: 1
    Tag:
      type: object
      description: Tag that can be attached to subscriptions and bookmarks
      required:
      - id
      - title
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the tag
        title:
          type: string
          description: Human-readable name of the tag, unique per user
        createdAt:
          type: string
          format: date-time
          description: Timestamp at which the tag was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp at which the tag was last modified
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- bearerAuth: []